+++ analysis/views/__init__.py
@@ -1,0 +1,3 @@
+from .list_view import VideoAnalysisListView
+from .detail_view import VideoAnalysisDetailView 
+from .ids_list_view import VideoAnalysisIdListView
