+++ analysis/views/list_view.py
@@ -1,8 +1,8 @@
 from rest_framework import generics, permissions
 from django_filters.rest_framework import DjangoFilterBackend
 from analysis.models import VideoAnalysis
-from .serializers import VideoAnalysisSerializer
-from .filters import VideoAnalysisFilter
+from ..serializers import VideoAnalysisSerializer
+from ..filters import VideoAnalysisFilter
 
 
 class VideoAnalysisListView(generics.ListCreateAPIView):
@@ -18,11 +18,3 @@ class VideoAnalysisListView(generics.ListCreateAPIView):
 
     def perform_create(self, serializer):
         serializer.save(user=self.request.user)
-
-
-class VideoAnalysisDetailView(generics.RetrieveUpdateDestroyAPIView):
-    serializer_class = VideoAnalysisSerializer
-    permission_classes = [permissions.IsAuthenticated]
-
-    def get_queryset(self):
-        return VideoAnalysis.objects.filter(user=self.request.user)
