+++ larvixon_site/settings.py
@@ -30,6 +30,7 @@ DEBUG = True
 
 ALLOWED_HOSTS: list[str] = []
 
+DEFAULT_PAGE_SIZE = 6
 
 # Application definition
 
@@ -194,6 +195,8 @@ REST_FRAMEWORK = {
     ],
     "DEFAULT_FILTER_BACKENDS": ["django_filters.rest_framework.DjangoFilterBackend"],
     "DEFAULT_SCHEMA_CLASS": "drf_spectacular.openapi.AutoSchema",
+    'DEFAULT_PAGINATION_CLASS': 'rest_framework.pagination.PageNumberPagination',
+    'PAGE_SIZE': DEFAULT_PAGE_SIZE
 }
 
 # Spectacular settings for API documentation
